Projects/Face_Mask_Detection/requirements.txt
tensorflow-cpu>=2.13.0
opencv-python>=4.8.0
numpy>=1.24.0
flask>=3.0.0
pillow>=10.0.0
